hotfix net separation
