+++ src/components/Sections/AppHeader.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { WarningTwoIcon } from "@chakra-ui/icons";
 import {
   Accordion,
